+++ modules/tinymce/src/core/main/ts/Deprecations.ts
@@ -13,12 +13,14 @@ import Tools from './api/util/Tools';
 const deprecatedSettings = (
   'autoresize_on_init,content_editable_state,convert_fonts_to_spans,inline_styles,padd_empty_with_br,block_elements,' +
   'boolean_attributes,editor_deselector,editor_selector,elements,file_browser_callback_types,filepicker_validator_handler,' +
-  'force_hex_style_colors,force_p_newlines,gecko_spellcheck,images_dataimg_filter,mode,move_caret_before_on_enter_elements,' +
+  'force_hex_style_colors,force_p_newlines,gecko_spellcheck,images_dataimg_filter,media_scripts,mode,move_caret_before_on_enter_elements,' +
   'non_empty_elements,self_closing_elements,short_ended_elements,special,spellchecker_select_languages,spellchecker_whitelist,' +
-  'tab_focus,table_responsive_width,text_block_elements,text_inline_elements,toolbar_drawer,types,validate,whitespace_elements'
+  'tab_focus,table_responsive_width,text_block_elements,text_inline_elements,toolbar_drawer,types,validate,whitespace_elements,' +
+  'paste_word_valid_elements,paste_retain_style_properties,paste_convert_word_fake_lists'
 ).split(',');
 
 const deprecatedPlugins = 'bbcode,colorpicker,contextmenu,fullpage,legacyoutput,spellchecker,textcolor'.split(',');
+const movedToPremiumPlugins = 'imagetools,toc'.split(',');
 
 const getDeprecatedSettings = (settings: RawEditorSettings): string[] => {
   const settingNames = Arr.filter(deprecatedSettings, (setting) => Obj.has(settings, setting));
@@ -32,7 +34,14 @@ const getDeprecatedSettings = (settings: RawEditorSettings): string[] => {
 
 const getDeprecatedPlugins = (settings: EditorSettings): string[] => {
   const plugins = Tools.makeMap(settings.plugins, ' ');
-  return Arr.sort(Arr.filter(deprecatedPlugins, (plugin) => Obj.has(plugins, plugin)));
+  const hasPlugin = (plugin: string) => Obj.has(plugins, plugin);
+  const pluginNames = [
+    ...Arr.filter(deprecatedPlugins, hasPlugin),
+    ...Arr.bind(movedToPremiumPlugins, (plugin) => {
+      return hasPlugin(plugin) ? [ `${plugin} (moving to premium)` ] : [];
+    })
+  ];
+  return Arr.sort(pluginNames);
 };
 
 const logDeprecationsWarning = (rawSettings: RawEditorSettings, finalSettings: EditorSettings): void => {