+++ modules/tinymce/src/core/main/ts/api/EventTypes.ts
@@ -5,7 +5,7 @@
  * For commercial licenses see https://www.tiny.cloud/
  */
 
-import { GetContentArgs, SetContentArgs } from '../content/ContentTypes';
+import { Content, GetContentArgs, SetContentArgs } from '../content/ContentTypes';
 import { FormatVars } from '../fmt/FormatTypes';
 import { RangeLikeObject } from '../selection/RangeTypes';
 import { UndoLevel } from '../undo/UndoManagerTypes';
@@ -18,8 +18,10 @@ import { InstanceApi } from './WindowManager';
 export interface ExecCommandEvent { command: string; ui?: boolean; value?: any }
 
 // TODO Figure out if these properties should be on the ContentArgs types
-export type GetContentEvent = GetContentArgs & { source_view?: boolean; selection?: boolean; save?: boolean };
-export type SetContentEvent = SetContentArgs & { source_view?: boolean; paste?: boolean; selection?: boolean };
+export type BeforeGetContentEvent = GetContentArgs & { source_view?: boolean; selection?: boolean; save?: boolean };
+export type GetContentEvent = BeforeGetContentEvent & { content: Content };
+export type BeforeSetContentEvent = SetContentArgs & { source_view?: boolean; paste?: boolean; selection?: boolean };
+export type SetContentEvent = BeforeSetContentEvent;
 
 export interface NewBlockEvent { newBlock: Element }
 
@@ -89,9 +91,9 @@ export interface EditorEventMap extends Omit<NativeEventMap, 'blur' | 'focus'> {
   'GetSelectionRange': { range: Range };
   'SetSelectionRange': SetSelectionRangeEvent;
   'AfterSetSelectionRange': SetSelectionRangeEvent;
-  'BeforeGetContent': GetContentEvent;
+  'BeforeGetContent': BeforeGetContentEvent;
   'GetContent': GetContentEvent;
-  'BeforeSetContent': SetContentEvent;
+  'BeforeSetContent': BeforeSetContentEvent;
   'SetContent': SetContentEvent;
   'LoadContent': { };
   'PreviewFormats': { };