+++ modules/tinymce/src/core/main/ts/content/GetContentImpl.ts
@@ -5,7 +5,7 @@
  * For commercial licenses see https://www.tiny.cloud/
  */
 
-import { Arr, Fun, Optional } from '@ephox/katamari';
+import { Fun, Optional, Type } from '@ephox/katamari';
 import { SugarElement } from '@ephox/sugar';
 
 import Editor from '../api/Editor';
@@ -16,6 +16,7 @@ import { isWsPreserveElement } from '../dom/ElementType';
 import * as TrimHtml from '../dom/TrimHtml';
 import * as Zwsp from '../text/Zwsp';
 import { Content, ContentFormat, GetContentArgs } from './ContentTypes';
+import { postProcessGetContent, preProcessGetContent } from './PrePostProcess';
 
 const trimEmptyContents = (editor: Editor, html: string): string => {
   const blockName = Settings.getForcedRootBlock(editor);
@@ -30,36 +31,30 @@ const setupArgs = (args: Partial<GetContentArgs>, format: ContentFormat): GetCon
   getInner: true
 });
 
-const getContentFromBody = (editor: Editor, args: GetContentArgs, format: ContentFormat, body: HTMLElement): Content => {
+const getContentFromBody = (editor: Editor, args: Partial<GetContentArgs>, format: ContentFormat, body: HTMLElement): Content => {
   const defaultedArgs = setupArgs(args, format);
-  const updatedArgs = args.no_events ? defaultedArgs : editor.fire('BeforeGetContent', defaultedArgs);
+  return preProcessGetContent(editor, defaultedArgs).fold(Fun.identity, (updatedArgs) => {
+    let content: Content;
+    if (updatedArgs.format === 'raw') {
+      content = Tools.trim(TrimHtml.trimExternal(editor.serializer, body.innerHTML));
+    } else if (updatedArgs.format === 'text') {
+      // return empty string for text format when editor is empty to avoid bogus elements being returned in content
+      content = editor.dom.isEmpty(body) ? '' : Zwsp.trim(body.innerText || body.textContent);
+    } else if (updatedArgs.format === 'tree') {
+      content = editor.serializer.serialize(body, updatedArgs);
+    } else {
+      content = trimEmptyContents(editor, editor.serializer.serialize(body, updatedArgs));
+    }
 
-  let content: string;
-  if (updatedArgs.format === 'raw') {
-    content = Tools.trim(TrimHtml.trimExternal(editor.serializer, body.innerHTML));
-  } else if (updatedArgs.format === 'text') {
-    // return empty string for text format when editor is empty to avoid bogus elements being returned in content
-    content = editor.dom.isEmpty(body) ? '' : Zwsp.trim(body.innerText || body.textContent);
-  } else if (updatedArgs.format === 'tree') {
-    content = editor.serializer.serialize(body, updatedArgs);
-  } else {
-    content = trimEmptyContents(editor, editor.serializer.serialize(body, updatedArgs));
-  }
+    // Trim if not using a whitespace preserve format/element
+    const shouldTrim = updatedArgs.format !== 'text' && !isWsPreserveElement(SugarElement.fromDom(body));
+    const trimmedContent = shouldTrim && Type.isString(content) ? Tools.trim(content) : content;
 
-  if (!Arr.contains([ 'text', 'tree' ], updatedArgs.format) && !isWsPreserveElement(SugarElement.fromDom(body))) {
-    updatedArgs.content = Tools.trim(content);
-  } else {
-    updatedArgs.content = content;
-  }
-
-  if (updatedArgs.no_events) {
-    return updatedArgs.content;
-  } else {
-    return editor.fire('GetContent', updatedArgs).content;
-  }
+    return postProcessGetContent(editor, trimmedContent, updatedArgs);
+  });
 };
 
-export const getContentInternal = (editor: Editor, args: GetContentArgs, format: ContentFormat): Content => Optional.from(editor.getBody())
+export const getContentInternal = (editor: Editor, args: Partial<GetContentArgs>, format: ContentFormat): Content => Optional.from(editor.getBody())
   .fold(
     Fun.constant(args.format === 'tree' ? new AstNode('body', 11) : ''),
     (body) => getContentFromBody(editor, args, format, body)