+++ modules/katamari/CHANGELOG.md
@@ -14,7 +14,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - The `Result` and `Optional` APIs are now thoroughly documented.
 - Added private `.tag` and `.value` properties to `Result`, to make them friendlier to view in the debugger or console.
 
-## 8.1.0 - TBD
+## 8.1.0 - 2021-10-11
 
 ### Added
 - Added new `Strings.toInt` and `Strings.toFloat` APIs to be able to parse a string and convert it to a number.