+++ modules/sugar/CHANGELOG.md
@@ -9,7 +9,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Changed
 - Upgraded to Katamari 9.0, which includes breaking changes to the `Optional` API used in this module.
 
-## 8.1.0 - TBD
+## 8.1.0 - 2021-10-11
 
 ### Added
 - Added new `parentElement` function to the `Traverse` API.